Fixed the three warnings on 64-Bit iPhones.

+++ TSCurrencyTextField/TSCurrencyTextField.m
@@ -74,7 +74,7 @@
 
 - (void) setAmount: (NSNumber *) amount
 {
-    NSString* amountString = [NSString stringWithFormat: @"%.*lf", _currencyNumberFormatter.maximumFractionDigits, amount.doubleValue];
+    NSString* amountString = [NSString stringWithFormat: @"%.*lf", (short)_currencyNumberFormatter.maximumFractionDigits, amount.doubleValue];
     [self setText: amountString];
 }
 
@@ -152,12 +152,12 @@
         return NO;
     }
     
-    int distanceFromEnd = textField.text.length - (range.location + range.length);
+    long int distanceFromEnd = textField.text.length - (range.location + range.length);
     
     NSString* changed = [textField.text stringByReplacingCharactersInRange: range withString: string];
     [textField setText: changed];
     
-    int pos = textField.text.length - distanceFromEnd;
+    long int pos = textField.text.length - distanceFromEnd;
     if ( pos >= 0 && pos <= textField.text.length )
     {
         [textField setCaratPosition: pos];
